# Conflicts:
#	erp-server/erp-server-oms/src/main/java/com/erp/server/oms/service/impl/SoB2cServiceImpl.java

+++ erp-sdk/erp-sdk-oms-mercado-local/src/main/java/com/sdk/oms/mercadolocal/dto/mercadolocal/order/PaymentsBean.java
@@ -106,7 +106,7 @@ public class PaymentsBean {
     @JsonProperty("site_id")
     private String siteId;
     @JsonProperty("payer_id")
-    private int payerId;
+    private long payerId;
     @JsonProperty("order_id")
     private long orderId;
     @JsonProperty("currency_id")
